+++ packages/turbo-stream/index.d.ts
@@ -12,6 +12,9 @@ export class TurboChannel extends Channel {
 export interface StartOptions {
   tagName: string
   channelClass: typeof TurboChannel
+  requestSocketIDHeader: bool | string
 }
+
+export const DEFAULT_SOCKET_HEADER: string
 
 export function start(cable: Cable, opts?: Partial<StartOptions>): void