+++ packages/turbo-stream/index.js
@@ -1,2 +1,2 @@
 export { TurboChannel } from './channel.js'
-export { start } from './start.js'
+export { start, DEFAULT_SOCKET_HEADER } from './start.js'